Clean Up of size_t everywhere.

+++ netcdf/netcdf_coards.h
@@ -156,7 +156,7 @@ template< class DATA_TYPE, int NUM_DIMS > class netcdf_coards :
         NcValues* values = variable->values() ;
         this->_data = new DATA_TYPE[N] ;
         for ( size_t n=0 ; n < N ; ++n ) {
-            this->_data[n] = (DATA_TYPE) values->as_double(n) ;
+            this->_data[n] = (DATA_TYPE) values->as_double((long) n) ;
             if ( ! isnan(missing) ) {
             	if ( this->_data[n] == missing ) {
             		this->_data[n] = filling ;
